+++ packages/preset-umi/templates/routePreloadOnLoad/preloadRouteFilesScp.js
